backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
google-auth==2.29.0
google-cloud-pubsub==2.21.1
google-cloud-secret-manager==2.20.0
google-cloud-monitoring==2.21.0
kubernetes==29.0.0
firebase-admin==6.5.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-multipart==0.0.9
